Refactored one tool factory I missed
