+++ core/src/main/java/com/vzome/core/kinds/HeptagonFieldApplication.java
@@ -16,7 +16,7 @@ import com.vzome.core.tools.LinearMapToolFactory;
 import com.vzome.core.tools.MirrorToolFactory;
 import com.vzome.core.tools.ScalingToolFactory;
 import com.vzome.core.tools.SymmetryToolFactory;
-import com.vzome.core.tools.TranslationTool;
+import com.vzome.core.tools.TranslationToolFactory;
 import com.vzome.core.viewing.AbstractShapes;
 import com.vzome.core.viewing.ExportedVEFShapes;
 import com.vzome.core.viewing.OctahedralShapes;
@@ -72,7 +72,7 @@ public class HeptagonFieldApplication extends DefaultFieldApplication
             case TRANSFORM:
                 result .add( new ScalingToolFactory( tools, this .symmetry ) );
                 result .add( new SymmetryToolFactory( tools, this .symmetry ) );
-                result .add( new TranslationTool.Factory( tools ) );
+                result .add( new TranslationToolFactory( tools ) );
                 break;
 
             case LINEAR_MAP:
@@ -101,7 +101,7 @@ public class HeptagonFieldApplication extends DefaultFieldApplication
                 result .add( new ScalingToolFactory( tools, this .symmetry ) .createPredefinedTool( "scale down" ) );
                 result .add( new ScalingToolFactory( tools, this .symmetry ) .createPredefinedTool( "scale up" ) );
                 result .add( new SymmetryToolFactory( tools, this .symmetry ) .createPredefinedTool( "rotate around red through origin" ) );
-                result .add( new TranslationTool.Factory( tools ) .createPredefinedTool( "b1 move along +X" ) );
+                result .add( new TranslationToolFactory( tools ) .createPredefinedTool( "b1 move along +X" ) );
                 break;
 
             default:
