+++ core/src/main/java/com/vzome/core/kinds/RootThreeFieldApplication.java
@@ -23,7 +23,7 @@ import com.vzome.core.tools.MirrorToolFactory;
 import com.vzome.core.tools.ProjectionToolFactory;
 import com.vzome.core.tools.ScalingToolFactory;
 import com.vzome.core.tools.SymmetryToolFactory;
-import com.vzome.core.tools.TranslationTool;
+import com.vzome.core.tools.TranslationToolFactory;
 import com.vzome.core.viewing.AbstractShapes;
 import com.vzome.core.viewing.DodecagonalShapes;
 import com.vzome.core.viewing.ExportedVEFShapes;
@@ -114,7 +114,7 @@ public class RootThreeFieldApplication extends DefaultFieldApplication
             case TRANSFORM:
                 result .add( new ScalingToolFactory( tools, this .symmetry ) );
                 result .add( new SymmetryToolFactory( tools, this .symmetry ) );
-                result .add( new TranslationTool.Factory( tools ) );
+                result .add( new TranslationToolFactory( tools ) );
                 result .add( new ProjectionToolFactory( tools ) );
                 break;
 
@@ -145,7 +145,7 @@ public class RootThreeFieldApplication extends DefaultFieldApplication
                 result .add( new ScalingToolFactory( tools, this .symmetry ) .createPredefinedTool( "scale down" ) );
                 result .add( new ScalingToolFactory( tools, this .symmetry ) .createPredefinedTool( "scale up" ) );
                 result .add( new SymmetryToolFactory( tools, this .symmetry ) .createPredefinedTool( "rotate around red through origin" ) );
-                result .add( new TranslationTool.Factory( tools ) .createPredefinedTool( "b1 move along +X" ) );
+                result .add( new TranslationToolFactory( tools ) .createPredefinedTool( "b1 move along +X" ) );
                 break;
 
             default:
