+++ core/src/main/java/com/vzome/fields/sqrtphi/SqrtPhiFieldApplication.java
@@ -28,7 +28,7 @@ import com.vzome.core.tools.LinearMapToolFactory;
 import com.vzome.core.tools.MirrorToolFactory;
 import com.vzome.core.tools.ScalingToolFactory;
 import com.vzome.core.tools.SymmetryToolFactory;
-import com.vzome.core.tools.TranslationTool;
+import com.vzome.core.tools.TranslationToolFactory;
 import com.vzome.core.viewing.AbstractShapes;
 import com.vzome.core.viewing.ExportedVEFShapes;
 import com.vzome.core.viewing.OctahedralShapes;
@@ -135,7 +135,7 @@ public class SqrtPhiFieldApplication extends DefaultFieldApplication
 			case TRANSFORM:
 				result .add( new ScalingToolFactory( tools, pentaSymm ) );
 				result .add( new SymmetryToolFactory( tools, pentaSymm ) );
-				result .add( new TranslationTool.Factory( tools ) );
+				result .add( new TranslationToolFactory( tools ) );
 				break;
 
 			case LINEAR_MAP:
